Cleanup base_functions.py
* Add downloader/downloader_utils.py and move there allowed_filetype, check_free_space
  and is_4xx_client_error since they used only in *downloaders.py
* Move regex_links and write_last_post_file to modules where they are used
* Also deduplicate basic_auth from *downloaders.py
