+++ cyberdrop_dl/downloader/old_downloaders.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import asyncio
 import logging
 import multiprocessing
-from base64 import b64encode
 from functools import wraps
 from http import HTTPStatus
 from random import gauss
@@ -14,11 +13,8 @@ from tqdm import tqdm
 from yarl import URL
 
 from cyberdrop_dl.base_functions.base_functions import (
-    allowed_filetype,
-    check_free_space,
     clear,
     get_db_path,
-    is_4xx_client_error,
     log,
     logger,
 )
@@ -28,10 +24,12 @@ from cyberdrop_dl.base_functions.sql_helper import SQLHelper
 from cyberdrop_dl.client.client import Client, DownloadSession
 from cyberdrop_dl.scraper.Scraper import ScrapeMapper
 
-
-async def basic_auth(username, password):
-    token = b64encode(f"{username}:{password}".encode('utf-8')).decode("ascii")
-    return f'Basic {token}'
+from .downloader_utils import (
+    allowed_filetype,
+    basic_auth,
+    check_free_space,
+    is_4xx_client_error,
+)
 
 
 def retry(f):
